Split test suites by types of fields
